requirements.txt

Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
django-filter>=24.2
python-dotenv>=1.0
django-ckeditor>=6.3.1
